extract plugin api to publishable module
