+++ plugin/build.gradle.kts
@@ -1,0 +1,51 @@
+plugins {
+    id("com.android.library")
+    kotlin("android")
+    `maven-publish`
+}
+
+android {
+    namespace = "io.github.revenge.plugin"
+    compileSdk = 35
+
+    defaultConfig {
+        minSdk = 24
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_1_8
+        targetCompatibility = JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+}
+
+dependencies {
+    compileOnly("com.facebook.react:react-android:0.71.8")
+}
+
+publishing {
+    repositories {
+        maven {
+            name = "GitHubPackages"
+            url = uri("https://maven.pkg.github.com/revenge-mod/revenge-xposed")
+            credentials {
+                username = System.getenv("GITHUB_ACTOR")
+                password = System.getenv("GITHUB_TOKEN")
+            }
+        }
+    }
+
+    publications {
+        create<MavenPublication>("gpr") {
+            afterEvaluate {
+                from(components["release"])
+            }
+
+            groupId = "io.github.revenge"
+            artifactId = "plugin"
+            version = version
+        }
+    }
+}
